sdk/state: make safe to ingest txs out of order (#319)
Make it safe to ingest transactions from the different escrow accounts out of order, specifically in regards to the balances.

This is a follow up to #312. To make ingestion more resilient. To prevent the channel from accidentally ingesting old state and moving the balance cached within the channel back to a previous value. This can happen because updating balances uses the snapshot of the account balance in the result meta.

As it happens this should also bring the ingestion to the point where transactions for the two escrow accounts can be ingested out of order, which should allow us to stream payments for only those two accounts again. I'll open a follow up PR to do that.