sdk/state: avoid rehashing txs and remove cmp usage (#328)
Avoid rehashing transactions.

We rehash transactions during signing and verification. It's pretty expensive because full transaction has to be serialized. We hold hashes of the transactions next to the transactions and so there's little value in rehashing.

I didn't add any benchmark tests because even without benchmarks this is a pretty clear win that we wouldn't recalculate the hash when it is literally sitting next to the transaction in the same struct.