sdk/state: prevent opens once open begins (#334)
Don't allow a new open to be proposed if an open has already been proposed.

The logic for validating whether it is safe to propose a new open checks whether the open is complete and has all signatures, but this would make it possible for someone to propose an open, share it with a counterparty, then propose another open. That second proposal would discard the first open that was proposed. This could be unsafe because the other participant would hold transactions that could be valid that the proposer no longer has any record.