sdk/state: rename some functions for clarity (#335)
Rename any `isEmpty` functions to `Empty`. Rename any `isFull` functions to `HasAllSignatures`.

When I was making the change for #334 it wasn't obvious to me what the `isFull` function meant. This has happened to me before when I've forgotten what it meant and I had to go look at it to check. I think by renaming it to `HasAllSignatures` we'll make the intent of the function clearer. I also think the function can be exported. I also renamed `isEmpty` to `Empty` because I think it could equally be exported.