sdk/state: allow proposing payments with zero amounts (#343)
Allow proposing payments with zero amounts.

The SDK prevents proposing payments that are zero amounts, but from best I can tell allows the receiver to accept a zero amount. It's not really necessary that we prevent zero payments and they can be a useful primitive for generating a new iteration for the current agreement.

This would make merging channels easier, and implementing turn taking buffered channels easier also.