sdk/agent: keep amounts as int64 (#345)
Use int64 for amounts.

The agent currently has amounts be strings. This is inconvenient for middlelayer code, such as the code required for buffering payments, that needs to wrap an agent and sit between application code and the agent. We use int64s in most place and it isn't a heavy lift for consumers of this package to use them either. For the sake of simplicity I think this is worth it.