examples/bufferedbenchmark: variety of improvements (#354)
A variety of changes to the bufferedbenchmark:
- Added a max queue size that limits how big the queue can get.
- Added a wait function that allows the calling code to wait until the queue is empty and idle.
- Added cli parameters for controlling the max queue size and the test size.
- Changed the internal to use the same patterns employed in stellar/go#3980

Adding a max queue size lets us run benchmarks with a variety of queue sizes to see how that impacts performance. Large queues use more memory, which affects application performance, so testing this is valuable.

Having a state where the queue is full requires that we have a way to find out when it is available again. Adding the wait functionality required restructuring the internals of how the signaling worked, and I used the same pattern that was used in stellar/go#3980 because I think it was easier to extend.