agent/bufferedagent: add stats and rename queue to buffer (#359)
Add stats. Rename the use of the terms queue and settlements to buffer.

To reduce the terms being used in multiple places to refer to the same thing. It was beginning to get confusing. When I was adding the stats it became clear that the existing use of multiple terms was a problem.