examples/console: make account creation more reliable (#365)
Retry account creation at the start of the console example.

To make account creation at the start of the console example more reliable. Creating the escrow account fails because we do it immediately after creating the other account and for some reason Core or Horizon reject it if the following transaction is delivered too soon.