sdk/agent: fix bug where memo gets dropped on retry (#367)
Include the memo in the retry payment that happens when an account is underfunded.

When making a payment if the account is underfunded we update its balance then try again. The retry should use the memo provided, but it isn't right now. This was causing problems for the bufferedagent, because the bufferedagent place information about the buffered payments in the memo, but the memo was being dropped on retries.