examples/console: use terminology agreement for each agreement (#392)
Use "agreement" to refer to each off-network payment that contains the buffered payments.

It is confusing talking about two types of payments, the aggregate payment and the buffered payments. It's simpler and easier to understand if in the buffered payments demo we describe the channel payment that is part of the close agreement as an agreement.