sdk/state, sdk/agent: finalize payment (#406)
Add a finalize payment function that a payment proposer can use to finalize a payment when the confirmer sends back their signatures.

Today this functionality is provided by the ConfirmPayment function that is used by both the proposer and the confirmer. Because it is used by both it accepts a full close agreement envelope and does a lot of validation of the contents. The proposer gets back the signatures from the confirmer and doesn't actually need to verify the agreement. They just need to attach the signatures and see if they're valid. That's what the finalize payment function does.

This change has no material performance improvement. I attempted the change to see if it improved performance and it did not in a meaningful way. Reflecting on it though, this does feel like it naturally makes sense and it simplifies some logic we have in the agent. When the agent is confirming the payment for a proposer it no longer needs to get the close agreement and attach the signatures.